refactor: consolidate interface definitions (issue #79)
Move scattered interface definitions to proper locations following
two-namespace architecture:

**WayfarerMobile.Core.Interfaces** (pure, no MAUI dependencies):
- Moved 9 interfaces from WayfarerMobile/Core/Interfaces/
- Extracted ITripSyncService, ITimelineSyncService from service files
- Created ITripNavigationService for TripNavigationService
- Added SyncEventArgs.cs and DownloadEventArgs.cs for shared event types

**WayfarerMobile.Interfaces** (MAUI/Mapsui dependent):
- Moved 8 layer/map interfaces with Mapsui dependencies
- Extracted IWikipediaService, IDownloadNotificationService from services
- Changed namespace to WayfarerMobile.Interfaces

**Cleanup**:
- Deleted redundant TripDownloadTypes.cs from test infrastructure
- Updated ~14 consumer files with correct namespace imports
- All 1692 tests pass
