+++ src/WayfarerMobile.Core/Interfaces/IApiClient.cs
