+++ src/WayfarerMobile.Core/Interfaces/IAppLifecycleService.cs
