+++ src/WayfarerMobile.Core/Interfaces/IAppLockService.cs
