+++ src/WayfarerMobile.Core/Interfaces/IExceptionHandlerService.cs
