+++ src/WayfarerMobile.Core/Interfaces/IGroupsService.cs
