+++ src/WayfarerMobile.Core/Interfaces/ILocationBridge.cs
