+++ src/WayfarerMobile.Core/Interfaces/IPermissionsService.cs
