+++ src/WayfarerMobile.Core/Interfaces/IWakeLockService.cs
