+++ src/WayfarerMobile/Services/GroupLayerService.cs
@@ -7,6 +7,7 @@ using Microsoft.Extensions.Logging;
 using NetTopologySuite.Geometries;
 using WayfarerMobile.Core.Interfaces;
 using WayfarerMobile.Core.Models;
+using WayfarerMobile.Interfaces;
 using WayfarerMobile.Shared.Utilities;
 using Color = Mapsui.Styles.Color;
 using Brush = Mapsui.Styles.Brush;
