+++ src/WayfarerMobile/Services/TimelineLayerService.cs
@@ -6,6 +6,7 @@ using Mapsui.Styles;
 using Microsoft.Extensions.Logging;
 using WayfarerMobile.Core.Interfaces;
 using WayfarerMobile.Core.Models;
+using WayfarerMobile.Interfaces;
 using Color = Mapsui.Styles.Color;
 using Brush = Mapsui.Styles.Brush;
 using Pen = Mapsui.Styles.Pen;
