+++ src/WayfarerMobile/Services/TimelineSyncService.cs
@@ -572,55 +572,3 @@ public class TimelineSyncService : ITimelineSyncService
 
     #endregion
 }
-
-/// <summary>
-/// Interface for timeline sync service.
-/// </summary>
-public interface ITimelineSyncService
-{
-    /// <summary>
-    /// Event raised when sync is rejected by server.
-    /// </summary>
-    event EventHandler<SyncFailureEventArgs>? SyncRejected;
-
-    /// <summary>
-    /// Event raised when sync is queued for offline retry.
-    /// </summary>
-    event EventHandler<SyncQueuedEventArgs>? SyncQueued;
-
-    /// <summary>
-    /// Event raised when sync completes successfully.
-    /// </summary>
-    event EventHandler<SyncSuccessEventArgs>? SyncCompleted;
-
-    /// <summary>
-    /// Update a timeline location with optimistic UI pattern.
-    /// </summary>
-    Task UpdateLocationAsync(
-        int locationId,
-        double? latitude = null,
-        double? longitude = null,
-        DateTime? localTimestamp = null,
-        string? notes = null,
-        bool includeNotes = false);
-
-    /// <summary>
-    /// Delete a timeline location with optimistic UI pattern.
-    /// </summary>
-    Task DeleteLocationAsync(int locationId);
-
-    /// <summary>
-    /// Process pending mutations when online.
-    /// </summary>
-    Task ProcessPendingMutationsAsync();
-
-    /// <summary>
-    /// Get count of pending mutations.
-    /// </summary>
-    Task<int> GetPendingCountAsync();
-
-    /// <summary>
-    /// Clear rejected mutations.
-    /// </summary>
-    Task ClearRejectedMutationsAsync();
-}
