+++ src/WayfarerMobile/Services/TripSyncService.cs
@@ -1801,260 +1801,3 @@ public class TripSyncService : ITripSyncService
 
     #endregion
 }
-
-/// <summary>
-/// Interface for trip sync service.
-/// </summary>
-public interface ITripSyncService
-{
-    /// <summary>
-    /// Event raised when sync is rejected by server.
-    /// </summary>
-    event EventHandler<SyncFailureEventArgs>? SyncRejected;
-
-    /// <summary>
-    /// Event raised when sync is queued for offline retry.
-    /// </summary>
-    event EventHandler<SyncQueuedEventArgs>? SyncQueued;
-
-    /// <summary>
-    /// Event raised when sync completes successfully.
-    /// </summary>
-    event EventHandler<SyncSuccessEventArgs>? SyncCompleted;
-
-    /// <summary>
-    /// Event raised when a create operation completes with server ID.
-    /// </summary>
-    event EventHandler<EntityCreatedEventArgs>? EntityCreated;
-
-    #region Place Operations
-
-    /// <summary>
-    /// Create a new place with optimistic UI pattern.
-    /// </summary>
-    Task<Guid> CreatePlaceAsync(
-        Guid tripId,
-        Guid? regionId,
-        string name,
-        double latitude,
-        double longitude,
-        string? notes = null,
-        string? iconName = null,
-        string? markerColor = null,
-        int? displayOrder = null);
-
-    /// <summary>
-    /// Update a place with optimistic UI pattern.
-    /// </summary>
-    Task UpdatePlaceAsync(
-        Guid placeId,
-        Guid tripId,
-        string? name = null,
-        double? latitude = null,
-        double? longitude = null,
-        string? notes = null,
-        bool includeNotes = false,
-        string? iconName = null,
-        string? markerColor = null,
-        int? displayOrder = null);
-
-    /// <summary>
-    /// Delete a place with optimistic UI pattern.
-    /// </summary>
-    Task DeletePlaceAsync(Guid placeId, Guid tripId);
-
-    #endregion
-
-    #region Region Operations
-
-    /// <summary>
-    /// Create a new region with optimistic UI pattern.
-    /// </summary>
-    Task<Guid> CreateRegionAsync(
-        Guid tripId,
-        string name,
-        string? notes = null,
-        string? coverImageUrl = null,
-        double? centerLatitude = null,
-        double? centerLongitude = null,
-        int? displayOrder = null);
-
-    /// <summary>
-    /// Update a region with optimistic UI pattern.
-    /// </summary>
-    Task UpdateRegionAsync(
-        Guid regionId,
-        Guid tripId,
-        string? name = null,
-        string? notes = null,
-        bool includeNotes = false,
-        string? coverImageUrl = null,
-        double? centerLatitude = null,
-        double? centerLongitude = null,
-        int? displayOrder = null);
-
-    /// <summary>
-    /// Delete a region with optimistic UI pattern.
-    /// </summary>
-    Task DeleteRegionAsync(Guid regionId, Guid tripId);
-
-    #endregion
-
-    #region Trip Operations
-
-    /// <summary>
-    /// Update a trip's metadata (name, notes) with optimistic UI pattern.
-    /// </summary>
-    Task UpdateTripAsync(
-        Guid tripId,
-        string? name = null,
-        string? notes = null,
-        bool includeNotes = false);
-
-    #endregion
-
-    #region Segment Operations
-
-    /// <summary>
-    /// Update a segment's notes with optimistic UI pattern.
-    /// </summary>
-    Task UpdateSegmentNotesAsync(
-        Guid segmentId,
-        Guid tripId,
-        string? notes);
-
-    #endregion
-
-    #region Area Operations
-
-    /// <summary>
-    /// Update an area's (polygon) notes with optimistic UI pattern.
-    /// </summary>
-    Task UpdateAreaNotesAsync(
-        Guid tripId,
-        Guid areaId,
-        string? notes);
-
-    #endregion
-
-    /// <summary>
-    /// Process pending mutations when online.
-    /// </summary>
-    Task ProcessPendingMutationsAsync();
-
-    /// <summary>
-    /// Get count of pending mutations.
-    /// </summary>
-    Task<int> GetPendingCountAsync();
-
-    /// <summary>
-    /// Get count of failed mutations (exhausted retries or rejected).
-    /// </summary>
-    Task<int> GetFailedCountAsync();
-
-    /// <summary>
-    /// Clear rejected mutations.
-    /// </summary>
-    Task ClearRejectedMutationsAsync();
-
-    /// <summary>
-    /// Reset retry attempts for all failed mutations.
-    /// </summary>
-    Task ResetFailedMutationsAsync();
-
-    /// <summary>
-    /// Cancel all pending mutations (discard changes).
-    /// </summary>
-    Task CancelPendingMutationsAsync();
-
-    /// <summary>
-    /// Clears all pending mutations for a specific trip.
-    /// Call this when a trip is deleted.
-    /// </summary>
-    Task ClearPendingMutationsForTripAsync(Guid tripId);
-
-    /// <summary>
-    /// Gets count of failed mutations for a specific trip.
-    /// </summary>
-    Task<int> GetFailedCountForTripAsync(Guid tripId);
-
-    /// <summary>
-    /// Resets retry attempts for a specific trip's failed mutations.
-    /// </summary>
-    Task ResetFailedMutationsForTripAsync(Guid tripId);
-
-    /// <summary>
-    /// Cancels pending mutations for a specific trip (discards unsynced changes).
-    /// Also restores original values in offline tables.
-    /// </summary>
-    Task CancelPendingMutationsForTripAsync(Guid tripId);
-}
-
-/// <summary>
-/// Event args for sync failure.
-/// </summary>
-public class SyncFailureEventArgs : EventArgs
-{
-    /// <summary>
-    /// Gets or sets the entity ID that failed to sync.
-    /// </summary>
-    public Guid EntityId { get; set; }
-
-    /// <summary>
-    /// Gets or sets the error message.
-    /// </summary>
-    public string ErrorMessage { get; set; } = string.Empty;
-
-    /// <summary>
-    /// Gets or sets whether this was a client error (4xx).
-    /// </summary>
-    public bool IsClientError { get; set; }
-}
-
-/// <summary>
-/// Event args for sync queued.
-/// </summary>
-public class SyncQueuedEventArgs : EventArgs
-{
-    /// <summary>
-    /// Gets or sets the entity ID that was queued.
-    /// </summary>
-    public Guid EntityId { get; set; }
-
-    /// <summary>
-    /// Gets or sets the message.
-    /// </summary>
-    public string Message { get; set; } = string.Empty;
-}
-
-/// <summary>
-/// Event args for sync success.
-/// </summary>
-public class SyncSuccessEventArgs : EventArgs
-{
-    /// <summary>
-    /// Gets or sets the entity ID that synced successfully.
-    /// </summary>
-    public Guid EntityId { get; set; }
-}
-
-/// <summary>
-/// Event args for entity created with server-assigned ID.
-/// </summary>
-public class EntityCreatedEventArgs : EventArgs
-{
-    /// <summary>
-    /// Gets or sets the temporary client ID used before sync.
-    /// </summary>
-    public Guid TempClientId { get; set; }
-
-    /// <summary>
-    /// Gets or sets the server-assigned ID.
-    /// </summary>
-    public Guid ServerId { get; set; }
-
-    /// <summary>
-    /// Gets or sets the entity type (Place, Region).
-    /// </summary>
-    public string EntityType { get; set; } = string.Empty;
-}
