+++ src/WayfarerMobile/Shared/Controls/PlaceDetailsSheet.xaml.cs
@@ -532,7 +532,7 @@ public partial class PlaceDetailsSheet : ContentView
 
         try
         {
-            var wikipediaService = Application.Current?.Handler?.MauiContext?.Services.GetService<Services.IWikipediaService>();
+            var wikipediaService = Application.Current?.Handler?.MauiContext?.Services.GetService<Interfaces.IWikipediaService>();
             if (wikipediaService == null)
             {
                 await ShowAlertAsync("Error", "Wikipedia service not available.");
