+++ src/WayfarerMobile/ViewModels/TimelineViewModel.cs
@@ -13,6 +13,7 @@ using NetTopologySuite.Geometries;
 using WayfarerMobile.Core.Helpers;
 using WayfarerMobile.Core.Interfaces;
 using WayfarerMobile.Core.Models;
+using WayfarerMobile.Interfaces;
 using WayfarerMobile.Data.Services;
 using WayfarerMobile.Services;
 using WayfarerMobile.Shared.Controls;
